filesystem resource mandatory closure enforced

+++ src/main/java/org/wiztools/crypt/Encrypt.java
@@ -18,33 +18,52 @@ import java.io.FileNotFoundException;
 import java.util.ResourceBundle;
 
 public class Encrypt implements IProcess{
-	private final String ALGO = ResourceBundle.getBundle("wizcrypt").getString("algorithm");
-
-	private Cipher cipher;
-	private byte[] passKeyHash;
-
-	public void init(String keyStr) throws NoSuchAlgorithmException, InvalidKeyException, NoSuchPaddingException{
-		SecretKey key = new SecretKeySpec(keyStr.getBytes(), ALGO);
-		cipher = Cipher.getInstance(ALGO);
-		cipher.init(Cipher.ENCRYPT_MODE, key);
-		passKeyHash = PassHash.passHash(keyStr);
-	}
-
-	public void process(File file) throws IOException, FileNotFoundException{
-		File outFile = new File(
-				file.getCanonicalPath()+".wiz");
-		CipherInputStream cis = new CipherInputStream(
-					new FileInputStream(file), cipher);
-		FileOutputStream fos = new FileOutputStream(outFile);
-		// Write the hash in first 16 bytes
-		fos.write(passKeyHash);
-
-		int i = -1;
-		while((i=cis.read()) != -1){
-			fos.write((char)i);
-		}
-		fos.close();
-		cis.close();
-		file.delete();
-	}
+    private final String ALGO = ResourceBundle.getBundle("wizcrypt").getString("algorithm");
+    
+    private Cipher cipher;
+    private byte[] passKeyHash;
+    
+    public void init(String keyStr) throws NoSuchAlgorithmException, InvalidKeyException, NoSuchPaddingException{
+        SecretKey key = new SecretKeySpec(keyStr.getBytes(), ALGO);
+        cipher = Cipher.getInstance(ALGO);
+        cipher.init(Cipher.ENCRYPT_MODE, key);
+        passKeyHash = PassHash.passHash(keyStr);
+    }
+    
+    public void process(File file) throws IOException, FileNotFoundException{
+        FileOutputStream fos = null;
+        CipherInputStream cis = null;
+        boolean canDelete = false;
+        try{
+            File outFile = new File(file.getCanonicalPath()+".wiz");
+            cis = new CipherInputStream(new FileInputStream(file), cipher);
+            fos = new FileOutputStream(outFile);
+            // Write the hash in first 16 bytes
+            fos.write(passKeyHash);
+            
+            int i = -1;
+            while((i=cis.read()) != -1){
+                fos.write((char)i);
+            }
+            canDelete = true;
+        } finally{
+            try{
+                if(fos != null){
+                    fos.close();
+                }
+            } catch(IOException ioe){
+                // Log Exception
+            }
+            try{
+                if(cis != null){
+                    cis.close();
+                }
+            } catch(IOException ioe){
+                // Log Exception
+            }
+            if(canDelete){
+                file.delete();
+            }
+        }
+    }
 }
